sessions/session3/requirement.txt
pandas
psycopg2-binary
pytest
pytest-cov
python-dotenv
